Documenting the goroutine issue; - Moving tests to  folder
